First commit.
Where to start.

The purpose of `better_image_tag` is to serve as a drop in replacement
for the stock Ruby on Rails `image_tag` helper. In addition to the base
functionality, we would like to be able to do some additional work like:

1. Inline image/asset data directly into a view to speed up rendering.
2. Force our images to have alt tags (configurable per project)
3. Be able to infer the images' width and height if they are not
   provided to us in the options hash (this is useful for images that are
   not in the asset pipeline and change based on a dynamic resource. EG: a
   partner's logo)

... and many more features.

This first commit contains the first trio above, with some tests.