Provide `#lazy_load` method and fix `#with_size`
The `#lazy_load` method provides one half of the equation when it comes
to lazily loading images on a page. This will get us an image tag with:

1. src attribute set to use a transparent 1x1 gif.
2. data-src attribute set to the ultimate, final, rendered image (url or file)

The other half of the equation is to use a JS tool like `lazySizes`:

  http://afarkas.github.io/lazysizes/

Setting up lazysizes in your javascript bundle(s) will handle the heavy
lifting of displaying the deferred image location.

***

Lastly, the fix for the with_size method is to allow any provided height
and width attributes to supersede the calculated image sizes.