Add configurable cache mechanism + webp support
1. With configurable cache - we can use the initializer to tell the
internals of this gem to use Rails' cache, or not. This allows us to
handle relatively safe ways to turn things off if there are any issues
either with caching, or our cache store, or just want to test things out
without cache being a possible impediment.

2. By passing the request into the initializer of our image tag object,
we can inspect whether the browser supports the webp format. IF it does,
then we return the tag with the webp extension instead of the original
files. Of course this will work for the asset pipeline if the file
exists.

TODO: Create rake task that converts non-webp images to webp.