Pull out environment check in ImageTag
Based on the type of test being written by an implementing application
(view, request, unit, feature, etc) the behavior here should be
predictable and static. Otherwise there are side-effects that cause
things to break.

For example - tests that exercise the full stack _need_ to have the
`{ use_super: true }` hash passed up through to the helper method in order
to pass on to the super method.