Add rake task to find jpgs and convert to webp
As part of this gem we provide a means to use a webp instead of a jpg if
the requesting browser supports the file format. As a means to create
those webps automatically this rake task is provided to loop through all
jpgs in the default rails assets' images path, and convert them to webp
using imagemagick's `convert` utility.

To use this within a rails application, add

```ruby
require "better_image_tag"
```

... in your application's `Rakefile`, just placed _before_ the line
containing:

```ruby
Rails.application.load_tasks
```

To override the default Rails image asset path you can do that in the
gem's config like so:

```ruby
BetterImageTag.configure do |config|
  config.images_path = "/path/to/images"
end
```