Treat inline svg's different from binary image data
SVG's are, essentially, a bunch of XML that we can spit directly out to
the browser instead of encoding to base64 and wrapping them within the
context of an image tag.

Because of this, there's bit of a special branching in these cases. If
it's an SVG, and we're inlining, spit the entire contents of it to the
browser. If we're passing a width and height, inline those width and
height attributes into the parent svg tag.