Fix bug with picture tags
In the tests, these work fine.
In real world usage, the picture tag feature did not.

It was not passing `use_super: true` to the view context to trigger
usage of the default image tag.

The fix: Specifically pass `use_super: true` to the view context instead
of the private helper method