Allow passing `enabled: boolean` to #lazy_load
Why?
====

Because sometimes there are partials that will be placed at the top of a
page where we do NOT want to lazy load an asset. For those times it
would be helpful to pass a boolean to the partial based on where in the
view it will be used