Fix issue with picture tags getting wrong classnames
If you pass the lazyload method to what ends up being a picture tag, the
`--picture` suffix will be applied to that `lazyload` class instead of
any other class that was passed in.

This commit fixes the behavior. It adds that `--picture` to all _other_
provided classnames, and leaves `lazyload` alone.